# Improve performance of coverage tracking