Instead of looking up a filename in a hash table, cache the hash table pointer in the extension part of the op_array itself. This is possible since PHP 7.4.

For the following benchmark:
```php
<?php

pcov\start();

$sum = 0.0;
for ($i = 0; $i < 5000000; $i++) {
    $sum+=$i;
}

pcov\stop();
echo "hi $sum\n";
```

On an i7-4790:
```
Benchmark 1: php -d extension=$(realpath modules/pcov_old.so) test.php
  Time (mean ± σ):     219.6 ms ±   3.4 ms    [User: 214.3 ms, System: 4.2 ms]
  Range (min … max):   216.3 ms … 229.3 ms    13 runs

Benchmark 2: php -d extension=$(realpath modules/pcov.so) test.php
  Time (mean ± σ):     176.9 ms ±   0.6 ms    [User: 172.4 ms, System: 4.0 ms]
  Range (min … max):   176.1 ms … 178.0 ms    16 runs

Summary
  php -d extension=$(realpath modules/pcov.so) test.php ran
    1.24 ± 0.02 times faster than php -d extension=$(realpath modules/pcov_old.so) test.php
```