@krakjoe 2025-06-29
Good idea, very nice result :)

On Sun, 29 Jun 2025, 21:51 Niels Dossche, ***@***.***> wrote:

> Instead of looking up a filename in a hash table, cache the hash table
> pointer in the extension part of the op_array itself. This is possible
> since PHP 7.4.
>
> For the following benchmark:
>
> <?php
> pcov\start();
> $sum = 0.0;for ($i = 0; $i < 5000000; $i++) {
>     $sum+=$i;
> }
> pcov\stop();echo "hi $sum\n";
>
> On an i7-4790:
>
> Benchmark 1: php -d extension=$(realpath modules/pcov_old.so) test.php
>   Time (mean ± σ):     219.6 ms ±   3.4 ms    [User: 214.3 ms, System: 4.2 ms]
>   Range (min … max):   216.3 ms … 229.3 ms    13 runs
>
> Benchmark 2: php -d extension=$(realpath modules/pcov.so) test.php
>   Time (mean ± σ):     176.9 ms ±   0.6 ms    [User: 172.4 ms, System: 4.0 ms]
>   Range (min … max):   176.1 ms … 178.0 ms    16 runs
>
> Summary
>   php -d extension=$(realpath modules/pcov.so) test.php ran
>     1.24 ± 0.02 times faster than php -d extension=$(realpath modules/pcov_old.so) test.php
>
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/krakjoe/pcov/pull/123
> Commit Summary
>
>    - 632087d
>    <https://github.com/krakjoe/pcov/pull/123/commits/632087d6763bde42d33b429d6fe3e6b79e4e5138>
>    Improve performance of coverage tracking
>
> File Changes
>
> (1 file <https://github.com/krakjoe/pcov/pull/123/files>)
>
>    - *M* pcov.c
>    <https://github.com/krakjoe/pcov/pull/123/files#diff-b087107e80a8fa279b6a96f34cc65279ad1b664934a363966d8a63322e667be4>
>    (79)
>
> Patch Links:
>
>    - https://github.com/krakjoe/pcov/pull/123.patch
>    - https://github.com/krakjoe/pcov/pull/123.diff
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/krakjoe/pcov/pull/123>, or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AARB52WU2UA7D7UYY34TWDL3GA7UDAVCNFSM6AAAAACAMRVLQWVHI2DSMVQWIX3LMV43ASLTON2WKOZTGE4DMNJRGA3TSNI>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>