@nielsdos 2025-06-29
@staabm Maybe it's interesting if you give this a try on your benchmark that you used today? The benchmark I used is idealized (but self-contained).