@staabm 2025-06-29
Can I install this change with pecl? Or some other easy way to try it?