@staabm 2025-06-30
I tried running with this PR, but get a segfault:

```
➜  infection git:(master) vendor/bin/phpunit --coverage-html coverage_report
PHPUnit 11.5.25 by Sebastian Bergmann and contributors.

Runtime:       PHP 8.3.22 with PCOV 1.0.13-dev
Configuration: /Users/m.staab/Documents/GitHub/infection/phpunit.xml.dist
Random Seed:   1751267866

..[1]    93089 segmentation fault  vendor/bin/phpunit --coverage-html coverage_report

```